+++ lib/src/node.dart
@@ -13,13 +13,19 @@ import '../responsibility_chain.dart';
 ///
 /// The other way to implement the handler is to use [FunctionHandler]s (or utilize [FunctionalNode]
 /// objects directly).
-///
+@Deprecated(
+  'ResponsibilityNode and FunctionalNode handlers are replaced by '
+  'IResponsibilityNodeBase and IResponsibilityNode respectively and will be removed in 2.0.0',
+)
 abstract class ResponsibilityNode<R, A> {
   /// A method that is called when the node is responsible for the getting the chain result.
   ///
   /// This method should either return FutureOr<[ChainResult]<[R]>> or throw an exception.
   /// Throwing an exception will have the same effect as returning the [ChainResult.failure].
-  ///
+  @Deprecated(
+    'ResponsibilityNode and FunctionalNode handlers are replaced by '
+    'IResponsibilityNodeBase and IResponsibilityNode respectively and will be removed in 2.0.0',
+  )
   FutureOr<ChainResult<R>> handle(A args);
 
   const ResponsibilityNode();
@@ -34,22 +40,30 @@ abstract class ResponsibilityNode<R, A> {
 /// It is not necessary to create the [FunctionalNode] objects directly - the
 /// [ResponsibilityChain.funcNode] method will create them automatically from the given
 /// [FunctionHandler].
-///
+@Deprecated(
+  'ResponsibilityNode and FunctionalNode handlers are replaced by '
+  'IResponsibilityNodeBase and IResponsibilityNode respectively and will be removed in 2.0.0',
+)
 class FunctionalNode<R, A> extends ResponsibilityNode<R, A> {
   /// A [FunctionHandler] taking the argument of the type [A] and returning a [ChainResult]<[R]>.
   ///
   /// The function is also allowed to throw an exception, which have the same effect as returning
   /// the [ChainResult.failure].
-  ///
   final FunctionHandler<R, A> handler;
 
   /// Creates a [FunctionalNode] from [VoidFunctionHandler] which does not take any arguments.
-  ///
+  @Deprecated(
+    'ResponsibilityNode and FunctionalNode handlers are replaced by '
+    'IResponsibilityNodeBase and IResponsibilityNode respectively and will be removed in 2.0.0',
+  )
   FunctionalNode(VoidFunctionHandler<R> handler)
       : handler = ((args) => handler());
 
   /// Creates a [FunctionalNode] from [FunctionHandler], which takes an argument of the type [A].
-  ///
+  @Deprecated(
+    'ResponsibilityNode and FunctionalNode handlers are replaced by '
+    'IResponsibilityNodeBase and IResponsibilityNode respectively and will be removed in 2.0.0',
+  )
   const FunctionalNode.withArgs(this.handler);
 
   @override